server/pyproject.toml
[project]
name = "chatba-server"
version = "0.1.0"
description = "chat ba server"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "accelerate==0.27.2",
    "aiofiles==23.2.1",
    "aiohttp==3.9.3",
    "aiolimiter==1.1.0",
    "aiosignal==1.3.1",
    "altair==5.2.0",
    "annotated-types==0.6.0",
    "anyio==4.3.0",
    "async-timeout==4.0.3",
    "attrs==23.2.0",
    "bce-python-sdk==0.9.5",
    "certifi==2024.2.2",
    "chardet==5.2.0",
    "charset-normalizer==3.3.2",
    "clevercsv==0.8.2",
    "click==8.1.7",
    "cmake==3.28.3",
    "colorama==0.4.6",
    "contourpy==1.2.0",
    "cycler==0.12.1",
    "dataclasses-json==0.6.4",
    "dill==0.3.8",
    "distro==1.9.0",
    "exceptiongroup==1.2.0",
    "fastapi==0.110.0",
    "ffmpy==0.3.2",
    "filelock==3.13.1",
    "fonttools==4.49.0",
    "frozenlist==1.4.1",
    "fsspec==2024.2.0",
    "future==1.0.0",
    "gradio==4.21.0",
    "gradio-client==0.12.0",
    "greenlet==3.0.3",
    "grpcio==1.62.1",
    "grpcio-tools==1.62.1",
    "h11==0.14.0",
    "h2==4.1.0",
    "hpack==4.0.0",
    "httpcore==1.0.4",
    "httpx==0.27.0",
    "huggingface-hub==0.21.4",
    "hyperframe==6.0.1",
    "idna==3.6",
    "ijson==3.2.3",
    "importlib-resources==6.1.3",
    "jinja2==3.1.3",
    "jsonpatch==1.33",
    "jsonpointer==2.4",
    "jsonschema==4.21.1",
    "jsonschema-specifications==2023.12.1",
    "kiwisolver==1.4.5",
    "langchain==0.2.2",
    "langchain-community>=0.2.4",
    "langchain-core==0.2.4",
    "langchain-openai>=0.1.7",
    "lit==17.0.6",
    "markdown-it-py==3.0.0",
    "markdown2==2.4.13",
    "markupsafe==2.1.5",
    "marshmallow==3.21.1",
    "matplotlib==3.8.3",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "multidict==6.0.5",
    "multiprocess==0.70.16",
    "mypy-extensions==1.0.0",
    "networkx==3.2.1",
    "nh3==0.2.15",
    "numpy==1.26.4",
    "openai==1.25.0",
    "orjson==3.9.15",
    "packaging==23.2",
    "pandas==2.2.1",
    "peft==0.9.0",
    "pillow==10.2.0",
    "portalocker==2.8.2",
    "prompt-toolkit==3.0.43",
    "protobuf==4.25.3",
    "psutil==5.9.8",
    "pyarrow==15.0.1",
    "pycryptodome==3.20.0",
    "pydantic-core==2.16.3",
    "pydub==0.25.1",
    "pygments==2.17.2",
    "pyparsing==3.1.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.9",
    "pytz==2024.1",
    "pyyaml==6.0.1",
    "qdrant-client==1.8.0",
    "qianfan==0.3.4",
    "referencing==0.33.0",
    "regex==2023.12.25",
    "requests==2.31.0",
    "rich==13.7.1",
    "rpds-py==0.18.0",
    "ruff==0.3.2",
    "safetensors==0.4.2",
    "semantic-version==2.10.0",
    "sentencepiece==0.2.0",
    "shellingham==1.5.4",
    "shortuuid==1.0.12",
    "six==1.16.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.28",
    "sse-starlette==2.0.0",
    "starlette==0.36.3",
    "svgwrite==1.4.3",
    "sympy==1.12",
    "tenacity==8.2.3",
    "tiktoken==0.7.0",
    "tokenizers==0.15.2",
    "tomlkit==0.12.0",
    "toolz==0.12.1",
    "torch==2.0.1",
    "torchaudio==2.0.2",
    "torchvision==0.15.2",
    "tqdm==4.66.2",
    "transformers==4.38.2",
    "triton==2.0.0",
    "typer==0.9.0",
    "typing-extensions==4.10.0",
    "typing-inspect==0.9.0",
    "tzdata==2024.1",
    "urllib3==2.2.1",
    "uvicorn==0.28.0",
    "wavedrom==2.0.3.post3",
    "wcwidth==0.2.13",
    "websockets==11.0.3",
    "yarl==1.9.4",
    "zipp==3.17.0",
]
